requirements.txt
fastapi
uvicorn[standard]
docker
supabase
python-dotenv